Rename interaction helpers file
To avoid confusion with property helpers when importing
